requirements.txt
pandas~=2.2.3
streamlit~=1.41.1
plotly~=5.24.1
numpy~=2.2.1
sqlalchemy~=2.0.37
torch
smolagents~=1.7.0
openpyxl
python-dotenv~=1.0.1
pydantic-settings~=2.7.1
fpdf~=1.7.2
kaleido